# openapi/openapi.yml
openapi: 3.0.3
info:
  title: "piscon-portal-v2"
  version: "v0.0.0"
servers:
  - url: "http://localhost:8080/api"
    description: "Local server"
...
paths:
  /oauth2/code:
    get:
      operationId: "getOauth2Code"
      tags:
        - "oauth2"
      summary: "OAuth2認証コードの取得・traQの認可ページにリダイレクト"
      description: "OAuth2の認証に必要なCode Verifierを生成し、traQの認可ページにリダイレクトします。"
      responses:
        303:
          description: "traQの認可ページにリダイレクトします。"
          headers:
            Location:
              schema:
                type: "string"
                format: "uri"
                example: "https://q.trap.jp/oauth2/authorize?response_type=code&client_id=client_id&redirect_uri=http://localhost:8080/api/oauth2/callback&state=state&code_challenge=code_challenge&code_challenge_method=S256"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  responses:
...
    InternalServerError:
      description: "Internal Server Error"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Internal Server Error"
  securitySchemes:
    UserAuth:
      type: apiKey
      in: cookie
      name: piscon_session
      description: "部員である場合にアクセスできる"
    AdminAuth:
      type: apiKey
      in: cookie
      name: piscon_session
      description: "管理者である場合にアクセスできる"
    TeamAuth:
      type: apiKey
      in: cookie
      name: piscon_session
      description: "チームメンバーである場合にアクセスできる"